requirements.txt
pre_commit==4.2.0
channels==3.0.5
crispy-bootstrap5==2025.6
Django==5.2.4
django-crispy-forms==2.4
django-filter==25.2
python-dotenv
daphne
channels_redis
# Test tools
model-bakery==1.20.5
Faker==37.12.0
parameterized==0.9.0
#Debuging tools
django-debug-toolbar==6.1.0
django-silk==5.4.3
